wgengine/magicsock: dial derp without holding send lock
Dialing derp was blocking conn.Send, which bubbled up into
wireguard-go, which expects Send to behave like writing to
a typical socket with no blocking. The result fired the
watchdog timer in wgengine.

Instead, create the buffer channel for the derp server and
return it immediately, dialing on another goroutine.

Fixes #137

Signed-off-by: David Crawshaw <crawshaw@tailscale.com>