ipn: filter portlist to a short list of "interesting" ports.
Uploading a list of all listeners turned out to be hopelessly naive.
Windows has zillions of them under normal use (yikes, but that's
another story), and Chrome creates and destroys them constantly,
possibly related to webrtc.

Instead, we'll only return a list of the ports that we expect to expose
in the UI within the near future.

Signed-off-by: Avery Pennarun <apenwarr@tailscale.com>