version: revert the filepath change from earlier commit
f81233524fddeec450940af8dc1a0dd8841bf28c changed a use of package 'path' to 'filepath'.
Restore it back to 'path', with a comment.

Also, use the os.Executable-based fallback name in the case where the
binary itself doesn't have Go module information. That was overlooked in
the original code.