tailscaled.service: Allow read/write in /etc
`ProtectSystem=strict` makes more-or-less entire system read-only, but
tailscaled needs write access in /etc/ to create and edit resolv*.conf
files.

While a blanket read/write permission is not very secure, the
alternative would be to specifically note down the various potential
filenames tailscaled might try and write to, and I’m not sure if there
are more than the four currently shown on my system.

The combination of `ProtectSystem=strict` while undoing the protection
of /etc is still a big improvement towards sandboxing the daemon
compared to not having either.

Signed-off-by: Frederik “Freso” S. Olesen <freso.dk@gmail.com>