types/logger: add NoRateLimit
Using a simple function signature as our logger type is elegant,
but it has a downside: Cross-layer communication is difficult.
Sometimes you have a logging wrapper than needs to signal something
to a different logging wrapper that might or might not be present.

To date, we have worked around this by adding/recognizing magic
prefixes or substrings in the format string. Examples include
log verbosity and exemptions from rate limiting.

This commit is an experiment in how to accomplish that
in a more structured, extensible way.

The idea is to add "silent" args to the argument list,
and corresponding verbs to the format string.
These args (and verbs) format to the empty string,
so they are invisible to anyone who printfs to see the logged string.
However, layers that expect them can inspect the args directly
and modify their behavior accordingly.

This commit demos that approach with NoRateLimit,
which lets a logger opt out of rate limiting.
We could use this in magicsock to remove the
hard-coded list of exempt "no rate limit" strings.

One significant downside to this approach is that any
logf function that flattens the format and args
and re-emits them using "%s" will lose all silent args.
Such logf functions also break the rate limiter,
and are generally discouraged; document this.

Signed-off-by: Josh Bleecher Snyder <josh@tailscale.com>