types/logger: add NoRateLimit
Using a simple function signature as our logger type is elegant,
but it has a downside: Cross-layer communication is difficult.
Sometimes you have a logging wrapper than needs to signal something
to a different logging wrapper that might or might not be present.

To date, we have worked around this by adding/recognizing magic
prefixes or substrings in the format string. Examples include
log verbosity and exemptions from rate limiting.

This commit is an experiment in how to accomplish that
in a more structured, extensible way.

The idea is to bundle up the format string and args into a
type that prints itself identically to the original format string and args.
This are invisible to anyone who printfs to see the logged string.
However, layers that expect such types can inspect the args directly
and modify their behavior accordingly.

This commit demos that approach with NoRateLimit,
which lets a logger opt out of rate limiting.
We could use this in magicsock to remove the
hard-coded list of exempt "no rate limit" strings.

Signed-off-by: Josh Bleecher Snyder <josh@tailscale.com>